receivepack: error out if transaction exists without Praefect server
If a connection was proxied to us and contains information about an
ongoing transaction, Gitaly needs to know how to connect to the Praefect
server as otherwise it can't reach out to the transaction service. It's
thus an error case transaction information exists, but no information
about Praefect could be extracted. Let's thus raise an error in these
cases to make failures more obviously related.