Make --literal-pathspecs optional for LastCommitForPath
By default, we assume the pathspec is a glob spec, but there is now a
boolean field in the protobuf to add the --literal-pathspecs argument.
Changes to GitLab Rails will be needed to use this field to fix
https://gitlab.com/gitlab-org/gitaly/-/issues/2857.

This restores the changes in
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2271 that were
reverted in https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2282,
but now the boolean flag allows the caller decide which mode to
use. Apparently GitLab was using the ListCommitForPath parameter as a
glob spec when computing a hash of a pipeline cache key
(https://gitlab.com/gitlab-org/gitlab/-/issues/222356).