praefect: Make transaction node's role available to nodes
Currently, proxied Gitaly nodes are completely oblivious to the fact
whether they are treated as a primary or a secondary node by Praefect.
While this is a nice property to have, it doesn't really match with
reality where we may have to leave out some processing in case we're a
secondary node that gets written to. The most important example here is
Git hooks, which shouldn't get executed multiple times. Skipping them on
secondaries is thus a hard requirement.

Make available a given node's role as part of the transaction that
Praefect injects into the connection. Transactions are the only context
in which we need this information, as without a transaction the target
node can always be treated as primary.