testhelper: Introduce and use function to compare Protobuf messages
Protobuf messages shouldn't be compared using `require.Equals`, as it
isn't able to cope with their internals correctly. As a result, it will
label messages as different even though they are semantically the same.
While this is not yet a problem, it will become one as soon as we
migrate to protoc-gen-go v1.4 or higher.

So let's do the right thing now already to make a possible migration
easier in the future by introducing a new test helper that performs the
equality check for us.