Support literal path lookups in LastCommitsForTreeRequest
Directories that contained glob characters (e.g. `:wq`) would not be
viewable in GitLab because Gitaly attempted to use the path as a search
pattern instead of interpreting the tree as a literal pathname. Just as
we fixed this in
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2285 for
LastCommitForPath, we need to add support for passing the Git
`--literal-pathspecs` argument for LastCommitsForTreeRequest.

Closes https://gitlab.com/gitlab-org/gitaly/-/issues/2888