Makefile: Verify hash of Git binaries archive
When downloading the Git binaries archive from the CI job, we directly
unpack its contents into the target directory without first verifying we
actually got the expected contents. If somebody is able to replace
contents of the archive with something different, then this is a quick
way to obtain arbitrary command execution on any developers machine.

Fix this by verifying the checksum of the Git binary archive first.