Makefile: Simplify test-go target
The `test-go` Makefile target uses a rather complex structure with a
subshell, making it hard to read at first. The intent of the subshell is
to always run both `go test` and `go-junit-report` in succession, even
if `go test` failed, but to return an error in case `go test` failed.

As we already use a temporary file to keep track of the exit code, we
can just remove the subshell. Due to `|| echo $$? >${TEST_EXIT}`, `go
test` will always succeed anyway and not cause the target to fail.