Makefile: Use proper target for NOTICE generation
We currently use a phony target "notice-tmp" to generate the NOTICE
file. But as we're generating a real file, it doesn't make much sense to
use a phony target as opposed to a real target which is the actualy
generated NOTICE file.

Fix this by adding two normal targets to generate both the NOTICE file
in the source and build directories and have "notice" and
"notice-up-to-date" depend on these.