Refactoring of 'starter' package and gitaly server factory
Extending package 'starter' with 'Parse', 'Compose' funcs
to work with gitaly address format.
'GracefulStoppableServer' is a common interface that is why
it has been moved to 'starter' package where it is used.
'GitalyServerFactory' moved into 'server' package as
'bootstrap' is used by praefect as well.
Also it now supports multiple servers to be started for
listening and track all of them correctly.