transactions: Introduce structure to store voters
Currently, nodes are being tracked via a simple map that stores votes by
node names. As we're about to extend the voting mechanism to support
weighted voting, we'll need to store additional information for each
node, most importantly its weight. So let's introduce a new `Voter`
structure to prepare for this future.