Ensure Praefect replication queue tables exist
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/1865 introduced
gitaly_replication_* tables on 2020-03-10, but
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/1948 dropped the
gitaly_ prefix on 2020-03-26.

As a result, any instance that migrated Praefect between those two dates
would have an inconsistent state of the Praefect database. The only way
to recover would be to drop the database entirely. To avoid that, we
readd the replication queue tables if they don't exist to avoid
migration errors.

Closes https://gitlab.com/gitlab-org/gitaly/-/issues/2896