transactions: Introduce type for votes
Votes are currently a simple byte slice and as a result its internals
like e.g. the expected length are required to be known across the
implementation. Let's refactor the code to use a separate "vote" type
that maps to a fixed-size byte array to improve the situation and
encapsulate the code.