coordinator: Fix flaky test case
With the new "distributed reads" feature flag, reads now get distributed
to all nodes which are currently known as up-to-date. Picking a node
happens completely at random right now by first selecting all known
healthy nodes and then using a random index into the array to pick one.

While this strategy works, it causes one of our tests to be flaky as the
assumption is that when distributed reads are enabled, that we'll route
all read requests to the secondary. Let's fix this assumption by
re-running the director multiple times, asserting that it chose
different nodes and that for each chosen note, the rewritten storage
name matches the chosen node's storage.