Makefile: Indent conditional blocks
Makefiles are quite unflexible when it comes to indentation. Most
importantly, it's not allowed to indent anything by a tab character
that's not a rule body. Instead, one is allowed to use spaces though to
indent e.g. conditional blocks.

So let's indent our conditional blocks with spaces to help readability.