Makefile: Do not print needlessly verbose commands
Our current use of the "@" prefix to not have make print out the command
it's about to execute is kind of arbitrary right now. This commit tries
to change that and improve the situation by applying a set of rules:

    — Every command that doesn't produce output, but is central to the
      target in question should not be silenced. E.g. "go install" will
      not produce any output, but is usually central to the question in
      target.

    - Every command that does produce output and is central to the
      target in question should be silenced. E.g. while "go get" is
      usually central to the target in question, it already prints what
      exactly is happening and thus should be silenced.

    - Commands that are preparatory in nature for the actual command
      should be silenced. This includes e.g. calls like "mkdir" or "rm".

    - Commands generating user-facing output like "echo" should always
      be silenced.

Hopefully, this makes the Makefile a tad more consistent with regards to
when we silence things and when we don't.