Improve Gitaly connection error messages
Gitaly-ruby executes gRPC calls into gitaly server for a set
of methods. Each call could fail and the error message won't
give you visibility of the target endpoint it was issued for.
By updating gRPC lib to a newer version the address should be
included into the error as stated in the
https://github.com/grpc/grpc/pull/22868 that is included into
https://github.com/grpc/grpc/releases/tag/v1.30.0
There is no 1.30.0 version for the ruby client lib at
https://rubygems.org/gems/grpc, that is why the latest
1.30.0.pre1 version was added.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/2786