testhelper: Fix building Go modules in testdirectory
When executing `make test`, it may happen that we end up with a
`internal/testhelper/testdata/home/go` directory containing the Go build
infrastructure. The cause for this is `testhelper.ConfigureGit()`, which
will set up the `HOME` environment variable to point into that exact
directory. As some tests execute `go build` to build commands they
depend on, e.g. as done in `ConfigureGitalySSH()`, this will cause Go to
use the new home directory to build the modules. We tried to avoid this
exact issue in `ConfigureGit()` already by also setting the `GOCACHE`
variable to the real Go cache directory, but this will only help in case
all modules already exist in the cache.

Fix the issue by also setting the `GOPATH` variable to the real Go home
directory.