Makefile: Make dependency hashes overridable
While it's possible to override the version of some of our dependencies,
it's not currently possible to override the URLs or hashes of
corresponding downloaded archives. As a result, it's mostly impossible
to actually change any of the versions as the hashes are guaranteed to
mismatch.

Let's fix this by making both URLs as well as hashes configurable from
the command line.