Makefile: Allow overriding defaults via a config file
While it's possible to override quite some different things of our
Makefiles, doing so is burdensome as the developer needs to keep in mind
to always pass those parameters to any calls of make(1). So let's make
this configurable in a persistent way by including an optional
"config.mak" file that may contain options set up by the developer. All
the developer now needs to do is write all variables he wants to
override into that file and then he's good to go.