Praefect: replication jobs health ping
The replication process could take unpredictable amount of time.
And to distinct between stale jobs (the process is finished, but
the entry was not updated) and time-consuming jobs the health-ping
was introduced. It updates timestamp of the event in the database
so it can be used to identify if replication process is going on
for this event.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2873