connections: Create new connecton pool interface
We've got several locations which need to create GRPC connections to one
of our own servers, either to Gitaly or to Praefect. In order to not
consistently re-connect to the same nodes, existing callsites need to
make sure to use a connection pool for this. This is rather easy to do
as GRPC connections allow concurrent multiplexing without further issue,
but it's still duplicated logic.

So let's create a new connection pooling interface. The interface is
currently trivial: except for creation of a new pool via
`connections.NewPool()`, there's only the `Dial()` function which
transparently either creates a new connection or re-uses an already
existing one.