hooks: Make sure to hash stdin for transactions on Ruby secondaries
Recently, Gitaly nodes which take part in a transaction started to
become aware of whether they are primary or secondary nodes, which is
done so that secondaries know to skip execution of Git hooks. But as a
result, secondaries stopped reading standard input when using the Ruby
pre-receive hook. While this is mostly fine as there is no hook to
consume stdin, it does have two implications:

- Proxying to such secondaries may cause us to halt if the target
  doesn't read stdin as he's not accepting any messages.

- Reading standard input is mandatory on the client-side, too, as it
  will cause us to not update the hash we use to vote on the
  transaction. Thus, we'll always disagree with the primary, even though
  secondaries will now all agree with each other as they essentially
  vote on the hash of the empty string.

Fix both issues by reading all of standard input and discarding it
immediately on secondaries. Add several tests for Go- and Ruby-backed
pre-receive hooks to verify that reference transactions are invoked and
that they produce the expected hashes.