transactions: Expose info about voting outcomes
With the previous commit, we started to keep track about the outcome of
a transaction for each voter. Right now, this information is only
availbale internally and is thus not yet of much use. So let's change
the transaction manager's `CancelFunc` signature to return not only an
error, but also the map of voting outcomes by node.