Add field name for git.Flag
Go vet, editors, and other tooling will emit warnings when a struct
doens't have the fields names, as it falls back on ordering. That could
introduce bugs when new fields are added with the same type at the
beginning of a struct.

To reduce the number of warnings in my editor, I ran a `sed` script to
update the code.