coordinator: Schedule replication jobs for failing nodes
When using reference transactions, we eventually want to move to a world
where a subset of nodes may fail the transaction, but the majority of
nodes will still commit it. As a prerequisite for such a world, we need
to make sure to repair any node that fails in such a quorum-based
transaction. If we didn't, it would cause us to never automatically
repair nodes after the first error happened.

This commit wires up exactly that logic. As we now have info available
about all nodes that failed during the transaction when calling its
cleanup function, we can schedule replication jobs for all nodes that
failed during the transaction. As this can only work if the primary
actually succeeded during the transaction, we need to make sure to _not_
schedule any jobs if the primary is also part of the failed voters. This
essentially means we'll only ever schedule jobs for secondaries.

As noted above, this is done as a prerequisite for quorum-based voting.
As we're not using them yet, we'll only commit if all nodes agree. As a
result, there is no partial-failure scenario yet and thus the new logic
shouldn't be exercised at all. This will change in the future, though.