praefect: Let mock SmartHTTP server hook into transactions
With the upcoming change of checking whether transactions failed or not
inside the coordinator in order to schedule replication jobs, we need to
make sure that reference transactions are properly set up and that
they're getting executed. One part that doesn't yet do so is the ad-hoc
SmartHTTP server implementation used to test proxy writes, so let's wire
it up to avoid future breakage.

This is kind of a stop-gap measure that only works because we currently
know that each reference-transaction will actually get called. This will
change at a later point when we extend our usage of reference
transactions beyond git-receive-pack(1), but let's just not think about
that yet.