transactions: Improve error message when vote fails to reach quorum
The error message when a vote fails to reach quorum is currently the
generic "transaction vote failed", which in this context could mean a
lot. What it does want to say is that the vote failed to reach at any
majority and thus did not reach quorum, so let's rewrite the error
message to be more explicit here and return the actual and required
number of votes.