remote: Make `RunRemoteServiceServer()` helper public
The `runRemoteServiceServer()` test helper sets up a new GRPC server
with the "remote" service preregistered. We'll want to make use of this
function in our FetchInternalRemote tests -- while these tests are
located in the same directory, they use a different package to fix an
import cycle. So in order to use it, we need to make the symbol publicly
visible.

Let's do so. While at it, let's also use `NewServer()` instead of
creating the server structure ad-hoc and accept a set of server options.