Praefect: replication processing should omit unhealthy nodes
Replication should consider the state of the target nodes and
omit dequeueing and processing of replication events for unhealthy
nodes. Processing of replication events for unhealthy nodes could
lead to outdated repositories as all attempts to process the
event would be failed and event will be marked 'dead'. And the
repository will remain outdated until the next replication event
happen to it. By checking the health state before processing
the event we have more chances to apply them without issues, so
reduce potential data loss.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/2922