transactions: Record outcome of a transaction
Previous to implementing weighted voting, it was trivial to determine
the outcome for each voter: either all voters committed or all voters
failed. But that has changed now as one subset may fail while another
subset succeeds and we need to get that information across to other
subsystems so they can act on it, e.g. by repairing all nodes that have
failed the vote.

The earliest point where we can know the outcome for all nodes is when
cancelling a transaction. At that point, each voter is in one of the
following three states:

    - committed: The voter committed the transaction and is thus known
                 to have succeeded.

    - aborted: The voter aborted the transaction and is thus known to
               have failed.

    - undecided: The voter didn't decide on his vote yet. This may
                 happen e.g. because it didn't show up yet or because no
                 majority was reached yet. We can only decide what to do
                 with those voters on transaction cancellation, which is
                 to mark them as failed. This is also the reason why we
                 can only make info about succeeding/failed nodes
                 available at the time of transaction cancellation,
                 only.

So let's make this information available by storing each nodes current
state, which is one of the three described ones, and have the `cancel()`
function return a map storing for each node, whether it has succeeded or
not. Note that at this point in time, we don't make use of that
information yet.