repository: Fix SearchFiles tests if Git does not support PCRE2
One of our tests for the `SearchFilesByContent` RPC call verifies that
we can successfully use PCRE2-only regular expressions. But as we rely
on Git actually supporting PCRE2 for this to work, the test will fail on
all machines that make use of a Git version that's not linked against
PCRE2 at all.

Fix the issue by skipping the test if we detect Git doesn't support
PCRE2-style regular exrpessions.