operations: Start using transactions for UserCreateBranch
Even though the `UserCreateBranch` RPC operates locally by executing Git
commands in the target repository directly, it still makes sure to
execute the same set of hooks one would expect to be called when
executing a git-push(1). As such, it will also call the pre-receive hook
and thus allow us to easily implement support for transactions on the
Ruby side. In fact, all we need to do is to extract transaction and
Praefect server metadata injected by Praefect and make it available to
the called hook via its environment variables.

So let's wire this up for the `UserCreateBranch` RPC as a first test
balloon on Ruby's side. This commit also adjusts some of our tests to
set the transactions feature flag so we actually exercise this code.