metadata: Make transaction serialization functions internal
The functions to serialize and deserialize transactions to and from a
string currently have public visibility even though they are not used
externally, nor are they intended to. So let's make them an internal
implementation detail.

While we're at it, let's also rename `fromSerialized()` to
`transactionFromSerialized()` as the former one is too ambiguous.