transactions: Introduce per-service feature flags
We currently have a single feature flag that enables or disables use of
reference transactions as a whole. But as we're about to introduce more
and more users of transactions, it will become awkward to only be able
to completely disable transactions everywhere in case there's any error.

This is why we now introduce a new set of service-scoped feature flags.
In case the global `ReferenceTransactions` feature flag is enabled, it
now becomes possible to disable reference transactions for subsets of
Gitaly in case only those subsets prove to be faulty. As such, this new
set of flags is default-on to make them opt-out instead of opt-in.