ruby: Adjust transaction-related keys
With commit 1cbfcd74 (metadata: Make transaction serialization functions
internal, 2020-07-09), the environment and metadata keys for transaction
and Praefect server information have changed to improve consistency with
other existing variables. As this was done at the same time as hooking
the Ruby server into transactions, both changes crossed and thus the
Ruby server is still using the old variables. This wasn't noticed as
currently there is no way to observe failure in case we don't hook into
transactions, but will be noticed as soon as another change currently in
flight lands that takes care to schedule replication jobs on schedule.

So let's adjust the keys to fix the incompatibility.