Praefect: include PgBouncer in CI
Run 'praefect_sql_test' with a new docker image that includes
PgBouncer. All interaction with PostgreSQL database for this
job are going throw a PgBouncer.
The connection throw PgBouncer has some limitations that is
why `default_timeout_ms` configuration option was removed
and 'statement_timeout' is not set for the connection.
The version of the PostgreSQL was also upgraded to 11.8.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/2816