transactions: Split out subtransactions type
Currently, each transaction allows us to perform a single vote only.
This limitation works well enough as long as we use the pre-receive hook
for transactions, as we know it will only ever be executed once. But as
soon as we migrate to the reference-transaction hook, it may get
executed arbitrarily many times. As a result, we need to extend
transactions to support a sequence of votes.

As a preparatory step, this commit splits out the actual voting
mechanism from the transaction into a new "subtransaction" type. The
intent is that a transaction will start to be a container for multiple
subtransactions, where subtransactions are created ad-hoc as required.
This refactoring is not expected to have any user-visible impact.