hooks: Implement reference transaction hook server
With the reference-transaction hook having been accepted upstream and
being part of the upcoming Git v2.28, we can now finally start making
use of it for our own transactions. As it is going to eventually replace
voting logic in the pre-receive hook, all it currently needs to do is
vote on a given transaction if we're being passed data about an ongoing
transaction.