hooks: Accept context instead of stream when voting
Until now, the `voteOnTransaction` function was only used in the context
of the pre-receive hook server. As such, it also took as first parameter
the pre-receive hook server, making it unusable for any of the other
servers. As we're about to introduce the reference-transaction hook
server which also performs voting, this is a bit inconvenient as we
certainly don't want to duplicate provided functionality.

As the only thing we used the passed-in server for was to reuse its
context, let's just generalize the function signature to pass in a
context directly.